+++ src/components/real/power/battery.cpp
