+++ src/components/real/power/initialize_battery.cpp
