+++ src/components/real/power/initialize_pcu_initial_study.hpp
@@ -3,10 +3,10 @@
  * @brief Initialize function of PCU_InitialStudy
  */
 
-#ifndef S2E_COMPONENTS_POWER_INITIALIZE_PCU_INITIAL_STUDY_HPP_
-#define S2E_COMPONENTS_POWER_INITIALIZE_PCU_INITIAL_STUDY_HPP_
+#ifndef S2E_COMPONENTS_REAL_POWER_INITIALIZE_PCU_INITIAL_STUDY_HPP_
+#define S2E_COMPONENTS_REAL_POWER_INITIALIZE_PCU_INITIAL_STUDY_HPP_
 
-#include <components/power/pcu_initial_study.hpp>
+#include <components/real/power/pcu_initial_study.hpp>
 
 /*
  * @fn InitPCU_InitialStudy
@@ -21,4 +21,4 @@
 PCU_InitialStudy InitPCU_InitialStudy(ClockGenerator* clock_gen, int pcu_id, const std::string fname, const std::vector<SAP*> saps, BAT* bat,
                                       double compo_step_time);
 
-#endif  // S2E_COMPONENTS_POWER_INITIALIZE_PCU_INITIAL_STUDY_HPP_
+#endif  // S2E_COMPONENTS_REAL_POWER_INITIALIZE_PCU_INITIAL_STUDY_HPP_
