+++ src/components/real/power/initialize_solar_array_panel.cpp
