+++ src/components/real/power/initialize_solar_array_panel.hpp
@@ -3,10 +3,10 @@
  * @brief Initialize function of SAP (Solar Array Panel)
  */
 
-#ifndef S2E_COMPONENTS_POWER_INITIALIZE_SOLAR_ARRAY_PANEL_HPP_
-#define S2E_COMPONENTS_POWER_INITIALIZE_SOLAR_ARRAY_PANEL_HPP_
+#ifndef S2E_COMPONENTS_REAL_POWER_INITIALIZE_SOLAR_ARRAY_PANEL_HPP_
+#define S2E_COMPONENTS_REAL_POWER_INITIALIZE_SOLAR_ARRAY_PANEL_HPP_
 
-#include <components/power/solar_array_panel.hpp>
+#include <components/real/power/solar_array_panel.hpp>
 
 /*
  * @fn InitSAP
@@ -32,4 +32,4 @@ SAP InitSAP(ClockGenerator* clock_gen, int sap_id, const std::string fname, cons
  */
 SAP InitSAP(ClockGenerator* clock_gen, int sap_id, const std::string fname, const SRPEnvironment* srp, double compo_step_time);
 
-#endif  // S2E_COMPONENTS_POWER_INITIALIZE_SOLAR_ARRAY_PANEL_HPP_
+#endif  // S2E_COMPONENTS_REAL_POWER_INITIALIZE_SOLAR_ARRAY_PANEL_HPP_
