+++ src/components/real/power/solar_array_panel.hpp
@@ -3,15 +3,15 @@
  * @brief Component emulation of Solar Array Panel
  */
 
-#ifndef S2E_COMPONENTS_POWER_SOLAR_ARRAY_PANEL_HPP_
-#define S2E_COMPONENTS_POWER_SOLAR_ARRAY_PANEL_HPP_
+#ifndef S2E_COMPONENTS_REAL_POWER_SOLAR_ARRAY_PANEL_HPP_
+#define S2E_COMPONENTS_REAL_POWER_SOLAR_ARRAY_PANEL_HPP_
 
 #include <environment/local/local_celestial_information.hpp>
 #include <environment/local/solar_radiation_pressure_environment.hpp>
 #include <library/logger/loggable.hpp>
 #include <library/math/vector.hpp>
 
-#include "../base/component.hpp"
+#include "../../base/component.hpp"
 
 class SAP : public ComponentBase, public ILoggable {
  public:
@@ -129,4 +129,4 @@ class SAP : public ComponentBase, public ILoggable {
   void MainRoutine(int time_count) override;
 };
 
-#endif  // S2E_COMPONENTS_POWER_SOLAR_ARRAY_PANEL_HPP_
+#endif  // S2E_COMPONENTS_REAL_POWER_SOLAR_ARRAY_PANEL_HPP_
